ClauLib/requirements.txt
future==1.0.0
iso8601==2.1.0
numpy==2.3.3
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
pyserial==3.5
PyYAML==6.0.3
scipy==1.16.2
